[go] Widen the carry type to 64 bits artifically
For performance reasons.  Note that this is unsound, but having the
carry type be 8 bits rather than 1 bit was already unsound, and this is
no more unsound.  The narrowness of the carry type is not used in
practice in the code, though we should still someday include it in the
proofs.

+++ fiat-go/32/poly1305/poly1305.go
@@ -27,8 +27,8 @@
 //   balance = [0x7fffff6, 0x7fffffe, 0x7fffffe, 0x7fffffe, 0x7fffffe]
 package poly1305
 
-type uint1 uint8
-type int1 int8
+type uint1 uint64 // We use uint64 instead of a more narrow type for performance reasons; see https://github.com/mit-plv/fiat-crypto/pull/1006#issuecomment-892625927
+type int1 int64 // We use uint64 instead of a more narrow type for performance reasons; see https://github.com/mit-plv/fiat-crypto/pull/1006#issuecomment-892625927
 
 // LooseFieldElement is a field element with loose bounds.
 //
